requirements.txt
# --- ANCHORS ---
langchain==0.3.7
langchain-community==0.3.7
docx2txt
langchain-core==0.3.17
langchain-text-splitters==0.3.2

# --- AI Sub-packages ---
# --- AI Sub-packages ---
langchain-openai==0.2.8
langchain-google-genai>=2.0.0
langchain-huggingface==0.1.2
langchain-unstructured==0.1.4
langgraph==0.2.55
sentence-transformers==3.3.1

# --- Fixed Versions for Compatibility ---
fastapi>=0.115.6
gradio>=5.0.0 
numpy>=1.22.4,<2.0.0
openai==1.54.3

# NOTE: The new SDK is google-genai. The warning suggests using it.
google-genai>=0.3.0
httpx>=0.28.1
# --- Database & Security ---
python-dotenv~=1.0.1
uvicorn>=0.34.0
sqlalchemy[asyncio]>=2.0.0
asyncpg==0.27.0
aiosqlite==0.20.0
psycopg2-binary
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==3.2.0
flask
flask_socketio
# --- Processing & Validation ---
pydantic[email]>=2.11.2,<3.0.0
pydantic-settings>=2.4.0
python-multipart
pygments==2.17.2
transformers
faster_whisper
ipywidgets==8.1.2
jupyter==1.0.0
langchain_experimental>=0.0.47
tiktoken
websockets>=13.0
# --- Vector & Unstructured ---
weaviate-client==4.10.2
langchain-weaviate==0.0.3
unstructured-client==0.25.9
unstructured[all-docs]==0.16.12
python-magic==0.4.27